Added user account change notifications
Whenever changes to a user account happen, we now catch them and
notify the user about the change. Of course, this is not a full
implementation but it provides the foundation for e.g. implementing
user confirmation on changing email addresses, or giving users a chance
to recover their accounts in case of malicious activities.

Signed-off-by: Daniel S. Reichenbach <daniel@kogitoapp.com>